Split out openapi crate into modules
