+++ openapi/Cargo.toml
@@ -11,3 +11,6 @@ reqwest = { version = "0.11", features = ["blocking"] }
 regex = "1"
 serde_json = "1"
 anyhow = "1.0.34"
+tracing = "0.1.29"
+log = "0.4.14"
+tracing-subscriber = "0.3.7"
